Implemented Tank Movement and Rotation
Set camera to be child object of tank when Game Starts so it can follow the tank as it moves.

Implemented receiving of movement input in TankView, logic of tank movement in TankController and setting of movementSpeed and rotationSpeed in TankModel.

Defined function to access TankModel instance via TankController.

TankView accesses movementSpeed and rotationSpeed in TankModel via TankController to call Move and Rotate functions defined in TankController.